Better name and profile
If it's supported by strftime, you can now put it in your name or profile, and it will automatically be changed to the current date and time. I don't advise including seconds or minutes, but if you want to get banned by botsin.space for excessive API calls, go for it.